scripts/requirements.txt
google-cloud-aiplatform>=1.38.0
Pillow>=9.0.0

